InsumoDAL creado y configurado

+++ Dominio.Core.Entities/Insumos.cs
@@ -28,6 +28,6 @@ namespace Dominio.Core.Entities
         [Display(Name = "Precio del insumo", Order = 4)]
         [Required(ErrorMessage = "El precio del insumo es necesario")]
         [RegularExpression(@"^\d{1,3}(\.\d{3})*(,\d+)?$/", ErrorMessage = "El precio debe ser valor numerico")]
-        public int in_dec_precio { get; set; }
+        public double in_dec_precio { get; set; }
     }
 }
